Move components, update success page logic
