+++ clients/payment/src/states/mandateUrl.ts
@@ -1,0 +1,6 @@
+import { atom, useAtom } from "react-atomic-state";
+
+const mandateUrl = atom<string | undefined>(undefined);
+
+export const setMandateUrl = mandateUrl.set;
+export const useMandateUrl = () => useAtom(mandateUrl);
